Fixed uint64 to int conversion issue by replacing all height types to uint64. (#275)
Renamed all "min" and "max" names to avoid collitions with the built-in functions.

+++ pkg/analysis/criteria/base_target_test.go
@@ -15,9 +15,9 @@ type baseTargetInfo struct {
 	node       string
 	ts         int64
 	v          string
-	h          int
+	h          uint64
 	sh         *proto.StateHash
-	baseTarget int
+	baseTarget uint64
 }
 
 func mkBaseTargetStatements(baseTargetInfo []baseTargetInfo) entities.NodeStatements {
