+++ pkg/messaging/pair/responses.go
@@ -31,7 +31,7 @@ func (nl *NodeStatementResponse) responseMarker() {}
 type NodeStatement struct {
 	URL       string              `json:"url"`
 	StateHash *proto.StateHash    `json:"statehash"`
-	Height    int                 `json:"height"`
+	Height    uint64              `json:"height"`
 	Status    entities.NodeStatus `json:"status"`
 	BlockID   *proto.BlockID      `json:"block_id"`
 	Generator *proto.WavesAddress `json:"generator"`
